New feature: Profiles
Users now have profiles that they can create and change at any time. More features, such as the ability to prefill the resume information with the profile information, will be added shortly. Additionally, the code to download a resume as a pdf has been removed and will be handled on the frontend instead