FormUser gone
All the forum data will also be handled in the base User class, mainly because of issues that changed how the authentication was registering.
Also, added a new null checker for @AuthenticationPrincipal parameters in controller methods, that throws if the parameter is null. This way, we won't have to manually do null checks each time we inject an authentication principal