Paging Resumes
Resumes are also returned as paginated results now, so when a user has many resumes we don't load them all at once. Additionally, both resumes and posts now use the modified time as a tiebreaker if the creation times are tied. Lastly, links to the previous and next pages are no longer directly sent to the client. Instead, the client is given all of the information to construct these links themselves when requested (when the user presses the next and previous methods)