Create user refactor (again)
Changed the newly introduced error data type to use generics instead, that way we can use it for multiple instances. Refactored the create user method to use it and also changed the process signup method from yesterday to use it as well. The resume and profile methods will be refactored the same way tomorrow