Database method refactoring
Begun refactoring the database access methods to manually rollback the transaction in memory if it fails (the database rolls itself back, so no problem there, but the state in memory will be corrupted). This will be completed at the same time as the previously stated refactor (on mostly the same methods), in the coming 1-2 days.