New factory methods
The optional error type now has new factory methods (that are also generic) to encapsulate away the details. The 3 factory methods cover the 4 use cases of an empty type, a type with a result and no error, and a type with a result and an error, and finally a type with just an error message and no result. These aren't being used yet, but will be with the next patch