Refactored the profile service
Changed the profile service methods to also use the new return type. Additionally, changed the resume service methods that modify a resume to return the modified version, instead of returning void, so that it can be sent back to the client. An analogous change will follow soon for the profile service methods as well