Some changes to the auth system
the refresh tokens are now stored on the client's browser by default, the "remember me" feature is unused for now, it will be properly reimplemented later.

Synced the expiration times of the refresh token cookies with that of the tokens themselves.

On logout, the refresh token cookie is deleted from the client (the next time they send a request to the server) and from the server's storage mechanisms

Finally, installed rabbitmq to act as a broker in the post and comment liking system