Updates to rate limiting
The rate limiting system no longer blocks proxy ip's as it is not very useful anyway. A new more comprehensive system will be built in the coming days.

The authentication system no longer sends refresh tokens to the client, because it is sent as a cookie instead.

Lastly, the request in the processLogin method has been refactored to use RequestEntity instead of just an HttpEntity