batch processing infra and rabbitmq
Created rabbitmq queues to enqueue like requests, and created item processors and writers to process them to our entities and write said entities to the database respectively. The item readers to read the objects in the first place will be developed soon.

Set up rabbitmq with 2 queues and 2 bindings with each of comment and post likes