New methods
Made methods to change comment and post likes easily, and created indexes on the comment and post classes to figure out if a user has liked a post more efficiently. The relations of PostLike and CommentLike now reference the email rather than the id, since that is the data that we get directly from the persistent authentication info. The client doesn't have the id of a user in the first place anyway