New entry forms, and other major changes
Added new forms for profile and resume section entries, and changed the existing section forms to contain a list of them instead. Also changed the methods that process said forms to account for this.

Additionally, fixed a bug in the resume update logic that caused references to old attributes to be changed, and made all service  methods take forms instead of individual form parameters