More form adoption
The resume forms also now contain forms of their fields, instead of each field of the forms directly. This allows for more abstraction, and saves us from modifying the code in 2 different places when adding a new form and/or a form field.

Additionaly, replaced a deprecated method with a newer one. (fromHttpUrl -> fromPath)