Changes to password rule system
Made the rules for the passwords a lot more relaxed by completely removing the maximum length requirement (in lieu of making another custom rule) and the maximum consecutive character and digit requirements. Also removed the ability to generate random passwords, enough of the complexity is removed now so that the user should be able to fix the mistakes themselves without many hassles.