SMS with twilio started
Prepared the codebase to move otp verification (by SMS only for now) entirely to twilio. The app will only handle the api calls to twilio and checking the status of the verification. Email OTP verification will likely be disabled for now, and will be readded with Twilio or potentially another provider.