More database simplification
The phone numbers are now just simply stored as Strings. This is obviously much easier to work with and eliminates the need for all of those conversions between the app, the client and the database. (Note to self: What were you thinking back then?)

All of the data in the forms, and the constructors have been adjusted. The phone-to-string converters (both the client and the database one) are gone.