Efficiently loading posts
Added the ability to load individual posts. It isn't fully done yet, because the comment entities might be merged back together. Updated the Comment and ForumPost entities so that most of their fields aren't sent to the client period with @JsonIgnore.

Made lazy loading explicit, and solved a problem with pagination returning most of the data anyway, making this feature almost obsolete. This is done by making the replies/comments @JsonIgnore'd, so that Jackson doesn't load them during pagination. This way, they are only explicitly sent to the client when the user explicitly loads a post.