Processing likes and persisting them in the database
Added logic for processing Like objects, and using them to update the like counts of the Post, Comment and CommentReply entities using Batch jobs.
The jobs don't process the like objects we already have processed for efficiency reasons.  Made a new writer and reader for the job. Configured the job to be run every 10 minutes at a fixed interval.