In memory rollback removed
Removed the logic of manually rolling back memory from all methods that did it. Instead, the relevant entities now have deep copy methods, the result of which acts like a pristine, unmodified copy and is returned to the client in case of failure (which is rare, since the constraints are all removed now). Also fixed some errors that came about as a result of implementing these changes and copy functions.

Lastly, replaced the uses of AntPathRequestMatcher (deprecated) with PatternPathRequestMatcher and changed the usage of DaoAuthentication provider to stop the deprecated use cases