remember me and better authentication
Made the login and sign up methods better, by only doing the log in logic if the check for a rate limit fails, and by making the sign up method automatically log the user in as well. Additionally, configured remember me services and the many beans that are necessary for it to function properly.

Finally, deleted the now useless remember me repository and service. Everything will be handled automatically by Spring