Update SecureAuthConfig.java
Disabled the formlogin setting. This is because that automatically injects a UsernamePasswordFilter on the login endpoint. This is a problem because:
1. We have 2 login endpoints.
2. The filter tries to go ahead and immediately do the authentication right then and there. However, this is problematic because the request hits the filter BEFORE the endpoint, and I don't have the time to verify the code the user entered to see if it is correct.

I could probably create my own implementation of that filter, but it would be a heavy modification. It is much easier to disable it.

OAuth2Login is disabled for the same reason, I might re-enable that later.