Improved timekeeping
Added the ability to support "Present" in experience times. Additionally added automatic sorting of experiences based on recency in the experience times.  Shortened the serialized versions of YearMonth objects, as well as updated the year month formatter to use the correct uuuu-MM format. The next thing to do is making users select (and be able to change) their timezones, and adjusting the values displayed to the user in their timezone.