Unlocked accounts are back
(Re) implemented locking and unlocking accounts. When  users don't log in for a long time, their accounts are considered deactivated, and they have to enter an OTP sent to their phone or email (whatever they entered on the initial login form). This also logs the user in, but it also marks their account as "active" which in reality is just setting the last logged in parameter back to the current time.

Additionally, changed the user details service to consider that the username parameter might also be a phone number as well. It can now handle that properly