Resume filtering added
Added a bunch of resume filtering parameters, as well as the code to build an elastic search query for resumes.

Removed all of the old specification code, as it is not needed anymore.

Finally, created a class that enables the easy creation of null-safe queries easily, abstracting away all of the logic of applying filters