Search token mechanism adjusted
For post requests to the relevant pages (i.e. initiating a search from a pre-saved configuration or a newly inputted form) the search token mechanism is entirely removed. It was only used to invalidate its old value, and it really doesn't make any sense to have it.

The token mechanism still exists on serving get requests to the same page, as intended. Lastly, since search tokens aren't expired programmatically anymore,  the search cache has a longer TTL and that timer now also resets when values are inspected.