More improvements
Added an endpoint to load the user's inbox data (only loads the pending friend requests that are received), and added an automatic job that runs every midnight that times out pending requests that are more than a week old.

Additionally fixed an issue that was caused by the likes processing job not being tolerant to any errors caused by optionals being empty.