Advancing on belonging enforcement checks
Added an instance of a StatementInspector to prepend a command to set a variable in postgresql before any of the sql statements go through. Added a context that holds the current id in a thread-safe manner, and then added an aspect to modify that context before and after every controller method execution.