More straightforwardness
The user context aspect has been removed, it is simply far too slow and too general. Additionally, the whole UserContext paradigm will also be removed very soon. It is simply too complex to add all of the RLS policies for every table in the database manually.  Even though the files are being pushed now, it will be changed very soon.