Additional checks for friend requests
Added belonging checks for friend requests, which simplified the methods of accepting and declining a friend request a lot more straightforward to read.  Changed the name of the method that actually accepts the friend requests to reflect that.

Another change is that friend requests no longer have statuses. Instead, they are deleted whenever they are accepted, rejected, or timed out.

Lastly, removed some useless enums and methods from the YearMonth utilities, dating way back to the start of the year.  Can't believe it took me this long to do that.