Messaging adjustments
The messsaging logic is improved slightly. There is now a setting (the utilities to modify said setting are also on the way) on each user's profile that allows the user to have the ability to not receive any new messages from anyone that they aren't friends with. The new messaging code now accounts for that.