Some API changes
Adjusted the URL's and placements (within controllers) of the API methods to be more consistent. Additionally, created a filter to automatically trim the trailing slash at the end of a request URL and to redirect to the correct address. Switched the CORS implementation to use Spring Security's recommended way instead of MVC