Some simplifications
The ExperienceType usage has been simplified. The value is now stored simply as a String, the enum itself is just being used to guide the generation of the possible values. The ExperienceType converter, bridge and the serializers have been removed.

The remaining 2 methods in the ObjectConverter class have been moved out into the respective classes that they were called from, which means that the ObjectConverter class has also been removed.