Encryption updated, timezone logic removed
The database encryptor class has been updated to use a TextEncryptor. (The functionality is still the same, it is just that the encryptor that is used already does what we did by ourselves manually).

Additionally, the logic to create and update a user timezone has been removed. Most of the timezone logic will be handled by the frontend.