Major configuration updates
Session management and logout configuration are simplified.

The resume serializer for Redis has been changed to use Java's default serialization mechanisms, which means that the explicit configuration for an ObjectMapper has been deleted.

The session management policy is now in one file instead of being scattered around in 3 different files.