More batch updates
The batch jobs no longer use job parameters. The parameters are now created and passed directly into the readers manually. This is more clear.

The Amqp readers are no longer restartable, and in fact they are now defined as @Bean components in the configuration classes that they are used in. The same applies to the writers, unless they have some special behavior that is not just persisting the entire chunk.

Lastly, the readers that do remain now also properly use the Instance type rather than LocalDateTime, I had overlooked this when doing the initial changes out of LocalDateTime.