Redis config changes
There is now a redis database in the cloud configured, and all redis configurations connect to that. Slightly adjusted the configurations to make setting the password, the url and the port possible (of course, they are set in our .env file).

The redis values have been removed from the application.yml file for now, they might be re-added as part of the different run profiles.