Comment logic changes
The comments no longer store the author usernames directly, as the data would be stale when the author changes their (forum) username.

Additionally, resolved the N+1 query issues when expanding a comment's replies, and while loading a post (and, by extension, its comments).  The user class also now holds a backup forum username field, in case the actual one is null or becomes null later. The COALESCE function is used to ensure the value sent to the client is never null.

Lastly, removed a now-redundant annotation (@ConfigurationPropertiesScan) from the main application class.