Lots of N+1 fixes
Improved the database performance by adjusting some queries (especially the ones used by the forum-related services) to be free of N+1 query problems. Made custom queries for some use cases.

The Inbox class has been removed, as it more often than not just served as an intermediary that did not really have a purpose (in fact, it made one query slower than it needed to be)

Lastly, for the versioning subsystem, made the queries that are used by it more performant by join fetching all other entities needed at load time, as well as fixing a small issue with the Experience versioning logic.