Cache system improvements
Centralized the cache key generation for the @Cacheable and @CacheEvict annotations.

Additionally, replaced all of the entity modification methods (which were originally designed to only change attributes that needed to change), with simply setting attributes on the objects directly. This is much more readable and maintainable.

The aspect that removed resumes from caches is gone. Most controller methods do that at the end anyway.

The next step tomorrow is to potentially simplify the resume and profile updating methods with one type, and to check if the changes today caused any issues.