Code duplication changes
Changed the resume post and put methods to go through the same underlying code, since they mostly do the same thing now (except for experiences).

Fixed an issue in the put controller regarding the key generation with caching.

Finally, deleted the experience, header and education repositories, since they are not needed and not used anymore.