Profile caching
Extended the caching mechanism to support profiles as well, in a manner that is very similar to how resume caching works.

All entities that are part of Resumes or UserProfiles now implement Serializable, a requirement for serialization to work properly.
This excludes the ForumPost, Comment and Chat entities with users for now, I may remove those from User and come up with methods to generate them at runtime, to reduce the sizes of cached objects.