passwords update
Passwords are no longer limited by custom rules, and instead whether they are pwned (via the haveIbeenpwned API) and a more heuristic score (via the zxcvbn library used by Dropbox), to be implemented in the near future.

To that end, all of the Passay-related password utilities have been deleted as they are not needed anymore.