Password improvements again
Incorporated the zxcvbn library for evaluating the strengths of passwords entered by the user.

The methods to evaluate the password and to do the entirety of the preliminary password checks are now stored in the UserService and the password service, in order to make sure individual classes don't become too big.